+++ contracts/nois-monitoring/src/error.rs
@@ -2,6 +2,7 @@ use cosmwasm_std::StdError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
+#[non_exhaustive]
 pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),