+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fix
+
+- nois-proxy: Embrace the use of empty `prices` lists to deactivate the proxy.
+  Turn a panic into an error when this happens.
+
 ## [0.13.0] - 2023-04-23
 
 ### Changed